# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FSP Talent API",
  "description": "\n**FSP Talent** — платформа подбора ИТ-специалистов с обратной механикой:\nработодатель сам находит кандидата в категоризированной базе и отправляет приглашение с вилкой ЗП.\n\n### Как авторизоваться в Swagger\n1. `POST /api/auth/register` → `POST /api/auth/verify-email` (в dev-режиме токен приходит в ответе регистрации).\n2. Кнопка **Authorize** вверху: username = email, password = пароль.\n\n### Формат ошибок\n`{\"code\": \"MACHINE_CODE\", \"message\": \"текст для пользователя\", \"detail\": \"...\"}`, для 422 ещё `fields: {поле: ошибка}`\n",
  "version": "1.3.0"
 },
 "paths": {
...
  "/api/invitations": {
   "post": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Работодатель: пригласить кандидата (вилка ЗП обязательна)",
    "operationId": "create_api_invitations_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Ошибка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Мои приглашения: работодатель — отправленные, кандидат — входящие",
    "operationId": "my_invitations_api_invitations_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "sent",
          "viewed",
          "accepted",
          "declined",
          "withdrawn"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InvitationOut"
         },
         "title": "Response My Invitations Api Invitations Get"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}": {
   "get": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Открыть приглашение (кандидату — автоматически статус viewed)",
    "operationId": "get_one_api_invitations__invitation_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}/accept": {
   "post": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Кандидат: принять — контакты станут видны работодателю",
    "operationId": "accept_api_invitations__invitation_id__accept_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/InvitationRespond"
         },
         {
          "type": "null"
         }
        ],
        "title": "Data"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}/decline": {
   "post": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Кандидат: отклонить",
    "operationId": "decline_api_invitations__invitation_id__decline_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/InvitationRespond"
         },
         {
          "type": "null"
         }
        ],
        "title": "Data"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}/answer": {
   "patch": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Кандидат: ответить одним запросом (status = accepted | rejected)",
    "description": "Повторный ответ на уже завершённое приглашение -> 409 INVALID_STATUS_TRANSITION.",
    "operationId": "answer_api_invitations__invitation_id__answer_patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InvitationAnswer"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}/withdraw": {
   "post": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Работодатель: отозвать приглашение",
    "operationId": "withdraw_api_invitations__invitation_id__withdraw_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/invitations/{invitation_id}/contact-access": {
   "put": {
    "tags": [
     "Приглашения (главная механика)"
    ],
    "summary": "Кандидат: закрыть свои контакты от компании (или открыть снова) после принятия",
    "description": "`{\"granted\": false}` — работодатель больше не видит контакты и полное имя; `{\"granted\": true}` — снова видит.",
    "operationId": "contact_access_api_invitations__invitation_id__contact_access_put",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "invitation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invitation Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ContactAccessIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InvitationOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CompanyShort": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "company_name": {
      "type": "string",
      "title": "Company Name"
     },
     "industry": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Industry"
     },
     "website": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Website"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "trust_level": {
      "type": "string",
      "title": "Trust Level",
      "default": "unverified"
     }
    },
    "type": "object",
    "required": [
     "id",
     "company_name"
    ],
    "title": "CompanyShort",
    "description": "То, что кандидат видит о компании в приглашении/вакансии."
   },
...
   "ContactAccessIn": {
    "properties": {
     "granted": {
      "type": "boolean",
      "title": "Granted",
      "description": "false — закрыть контакты от этой компании, true — открыть снова"
     }
    },
    "type": "object",
    "required": [
     "granted"
    ],
    "title": "ContactAccessIn"
   },
...
   "ErrorResponse": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "items": {},
        "type": "array"
       }
      ],
      "title": "Detail"
     },
     "fields": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "string"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fields"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "detail"
    ],
    "title": "ErrorResponse"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InvitationAnswer": {
    "properties": {
     "status": {
      "type": "string",
      "enum": [
       "accepted",
       "rejected",
       "declined"
      ],
      "title": "Status",
      "description": "rejected и declined — одно и то же"
     },
     "reply": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Reply"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "InvitationAnswer"
   },
   "InvitationCandidateShort": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_name",
     "category"
    ],
    "title": "InvitationCandidateShort"
   },
   "InvitationCreate": {
    "properties": {
     "salary_from": {
      "type": "integer",
      "maximum": 10000000.0,
      "exclusiveMinimum": 0.0,
      "title": "Salary From",
      "description": "ЗП от, ₽ в месяц",
      "examples": [
       200000
      ]
     },
     "salary_to": {
      "type": "integer",
      "maximum": 10000000.0,
      "exclusiveMinimum": 0.0,
      "title": "Salary To",
      "description": "ЗП до, ₽ в месяц",
      "examples": [
       280000
      ]
     },
     "salary_type": {
      "type": "string",
      "enum": [
       "gross",
       "net"
      ],
      "title": "Salary Type",
      "description": "gross — до вычета НДФЛ, net — «на руки». Подписываем явно, как просили организаторы",
      "default": "gross"
     },
     "candidate_id": {
      "type": "integer",
      "title": "Candidate Id"
     },
     "vacancy_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vacancy Id",
      "description": "Необязательно: можно пригласить без вакансии"
     },
     "position_title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Position Title",
      "description": "Необязательно: по умолчанию название вакансии или «Предложение о работе»",
      "examples": [
       "Middle Python-разработчик"
      ]
     },
     "message": {
      "type": "string",
      "maxLength": 5000,
      "minLength": 3,
      "title": "Message",
      "description": "Описание предложения",
      "examples": [
       "Добрый день! Мы строим платёжный сервис..."
      ]
     },
     "contact_method": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255,
        "minLength": 3
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Method",
      "description": "Как связаться. Не передан — берём «способ связи» из профиля компании",
      "examples": [
       "Telegram @hr_anna"
      ]
     }
    },
    "type": "object",
    "required": [
     "salary_from",
     "salary_to",
     "candidate_id",
     "message"
    ],
    "title": "InvitationCreate"
   },
   "InvitationOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "status": {
      "type": "string",
      "title": "Status",
      "description": "sent | viewed | accepted | declined | withdrawn. Для UI: sent и viewed = «Ожидает ответа»"
     },
     "candidate_id": {
      "type": "integer",
      "title": "Candidate Id"
     },
     "employer_id": {
      "type": "integer",
      "title": "Employer Id"
     },
     "position_title": {
      "type": "string",
      "title": "Position Title"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "salary_from": {
      "type": "integer",
      "title": "Salary From"
     },
     "salary_to": {
      "type": "integer",
      "title": "Salary To"
     },
     "salary_type": {
      "type": "string",
      "title": "Salary Type",
      "description": "gross — до вычета НДФЛ, net — на руки",
      "default": "gross"
     },
     "salary_type_label": {
      "type": "string",
      "title": "Salary Type Label",
      "default": "до вычета НДФЛ",
      "examples": [
       "до вычета НДФЛ",
       "на руки"
      ]
     },
     "currency": {
      "type": "string",
      "title": "Currency",
      "default": "RUB"
     },
     "contacts_revoked": {
      "type": "boolean",
      "title": "Contacts Revoked",
      "description": "Кандидат закрыл контакты после принятия",
      "default": false
     },
     "contact_method": {
      "type": "string",
      "title": "Contact Method"
     },
     "vacancy_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vacancy Id"
     },
     "company": {
      "$ref": "#/components/schemas/CompanyShort"
     },
     "candidate": {
      "$ref": "#/components/schemas/InvitationCandidateShort"
     },
     "match_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Match Score"
     },
     "match_reasons": {
      "items": {
       "$ref": "#/components/schemas/Reason"
      },
      "type": "array",
      "title": "Match Reasons"
     },
     "candidate_reply": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidate Reply"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "viewed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Viewed At"
     },
     "responded_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Responded At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "status",
     "candidate_id",
     "employer_id",
     "position_title",
     "message",
     "salary_from",
     "salary_to",
     "contact_method",
     "vacancy_id",
     "company",
     "candidate",
     "match_score",
     "match_reasons",
     "candidate_reply",
     "created_at",
     "viewed_at",
     "responded_at",
     "updated_at"
    ],
    "title": "InvitationOut"
   },
   "InvitationRespond": {
    "properties": {
     "reply": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Reply",
      "description": "Комментарий кандидата (необязательно)"
     }
    },
    "type": "object",
    "title": "InvitationRespond"
   },
...
   "Reason": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type",
      "description": "test | skills | text | fsp | strengths | activity | grade"
     },
     "label": {
      "type": "string",
      "title": "Label",
      "examples": [
       "Бэкенд · Middle: тест 87/100"
      ]
     },
     "positive": {
      "type": "boolean",
      "title": "Positive",
      "description": "True — подсвечивать зелёным"
     },
     "missing": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Missing",
      "description": "Для type=skills: каких навыков не хватает"
     }
    },
    "type": "object",
    "required": [
     "type",
     "label",
     "positive"
    ],
    "title": "Reason",
    "description": "Плашка-обоснование для UI: почему кандидат в выдаче."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/token"
     }
    }
   }
  }
 }
}